Use layers to draw minimap
